+++ src/runconf_ui/utils/shifter_config_tools/shifter_config_systems/component_extractor.py
@@ -1,4 +1,4 @@
-from runconf_ui.utils.shifter_config_tools.daq_system_readers.extractor_interfaces import SubsystemExtractor, SubsystemStatus
+from runconf_ui.utils.shifter_config_tools.shifter_config_systems.extractor_interfaces import SubsystemExtractor, SubsystemStatus
 import runconf_ui.interfaces.actions.actions as ca
 from runconf_ui.exceptions import CiderBadActionException
 from runconf_ui.utils.subsystem_status import SubsystemStatus
