+++ src/runconf_ui/utils/shifter_config_tools/shifter_config_systems/system_extractor.py
@@ -1,9 +1,9 @@
 from runconf_ui.interfaces.controller.daq_conf_wrapper import DaqConfigurationWrapper
-from runconf_ui.utils.shifter_config_tools.daq_system_readers.extractor_interfaces import (MultiItemExtractor,
+from runconf_ui.utils.shifter_config_tools.shifter_config_systems.extractor_interfaces import (MultiItemExtractor,
                                                                                            SubsystemExtractor,
                                                                                            SubsystemStatus)
-from runconf_ui.utils.shifter_config_tools.daq_system_readers.attribute_extractor import AttributeExtractor
-from runconf_ui.utils.shifter_config_tools.daq_system_readers.component_extractor import ComponentExtractor
+from runconf_ui.utils.shifter_config_tools.shifter_config_systems.attribute_extractor import AttributeExtractor
+from runconf_ui.utils.shifter_config_tools.shifter_config_systems.component_extractor import ComponentExtractor
 from runconf_ui.utils.subsystem_status import SubsystemStatus
 from runconf_ui.exceptions import CiderBadActionException
